Coding - Scope 25 Aspect enums as enum class
Converts 25 unscoped Aspect enums to enum class:
Aspect_Eye, Aspect_WidthOfLine, Aspect_GraphicsLibrary,
Aspect_TypeOfColorScaleData/Orientation/Position,
Aspect_XRActionType/GenericAction/TrackedDeviceRole,
Aspect_TypeOfDeflection/Marker/Line/Facing/StyleText/HighlightMethod/
DisplayText/Resize/Triedron, Aspect_FillMethod, Aspect_GradientFillMethod,
Aspect_InteriorStyle, Aspect_ColorSpace, Aspect_HatchStyle,
Aspect_XAtom, Aspect_GridDrawMode, Aspect_GridType.

Value names preserved; callers updated to scoped form.

Excluded because they are bitflag enums or integer-typed in practice:
- Aspect_PolygonOffsetMode (bitflags)
- Aspect_TypeOfTriedronPosition (bitflags)
- Aspect_VKeyBasic (Aspect_VKey typedef'd to unsigned int)

To keep OCCT_DUMP_FIELD_VALUE_NUMERICAL usable with the new enum
class fields, a generic `operator<<` template is added in
Standard_Dump.hxx that streams any scoped enum as its underlying
integer type.

Additional localized casts:
- array subscripts and sizes using enum values wrap with static_cast<int>()
- arithmetic on enum values (subtraction, addition) casts both sides
- `int` members storing enum values cast at initialization
- anonymous count-enums (Aspect_XRGenericAction_NB, etc.) cast their
  upper-bound enum class values

+++ src/Visualization/TKService/Graphic3d/Graphic3d_HatchStyle.hxx
@@ -35,7 +35,7 @@ public:
   //! Creates a new predefined hatch style with the given id in Aspect_HatchStyle enum.
   //! GPU memory for the pattern will not be allocated.
   Graphic3d_HatchStyle(const Aspect_HatchStyle theType)
-      : myHatchType(theType)
+      : myHatchType(static_cast<int>(theType))
   {
   }
